+++ Packages/ai.natml.vision.yolox/Changelog.md
@@ -1,3 +1,9 @@
+## 1.0.3
++ Added `YOLOXPredictor.Create` static method for fetching the model and creating the predictor.
++ Added `YOLOXPredictor.Detection` struct as the return type of the predictor.
++ Removed public `YOLOXPredictor` construcor. Use `YOLOXPredictor.Create` method instead.
++ Upgraded to NatML 1.1.2.
+
 ## 1.0.2
 + Upgraded to NatML 1.1.